requirements.txt
requests
beautifulsoup4
pandas
Flask
google-generativeai